experiments/pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "chromadb>=1.0.7",
    "huggingface-hub>=0.30.2",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-huggingface>=0.2.0",
    "langfuse>=2.60.1",
    "levenshtein>=0.27.1",
    "notebook>=7.3.2",
    "ollama>=0.4.7",
    "openai>=1.68.2",
    "openpyxl>=3.1.5",
    "pydantic>=2.10.6",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "ruff>=0.10.0",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "unidecode>=1.3.8",
]
